requirements.txt
# Microsoft Agent Framework
agent-framework-a2a==1.0.0b260319
agent-framework-ag-ui==1.0.0b260319
agent-framework-azure-ai==1.0.0rc5
agent-framework-core==1.0.0rc5
agent-framework-declarative==1.0.0b260319
agent-framework-devui==1.0.0b260319
agent-framework-mem0==1.0.0b260319
azure-ai-evaluation==1.14.0
azure-identity==1.25.3
azure-monitor-opentelemetry-exporter==1.0.0b49
opentelemetry-semantic-conventions-ai==0.4.13
# Requirements File
a2a-sdk==0.3.22
ag-ui-protocol==0.1.13
fastapi==0.128.0
fastmcp==2.14.4
httpx==0.28.1
openai==2.16.0
pydantic==2.12.5
python-dotenv==1.2.1
pytz==2025.2
pandas==2.3.3
uvicorn==0.40.0
uv